refactor(data): rename module of indexed datasets
