+++ datasets/__init__.py
@@ -5,8 +5,9 @@ The wrappers here include proper
 
 from .util import train_val_split, stratified_train_val_split
 from .path import dataset_path, DatapathMixin
-from .cache import IndexedImageFolder, IndexedDatasetFolder, IndexedImageDataset
+from .indexed import IndexedImageFolder, IndexedDatasetFolder, IndexedImageDataset
 from .subset import subset_dataset
 from .selfsuper import SelfsuperDataset, self_supervised
 from .vision import *
 from .image_folder_tar import ImageFolderTar
+from .cuda import CUDACached
